+++ example/lib/widgets/buttons/toggle_flashlight_button.dart
